+++ packages/core/src/index.ts
@@ -31,6 +31,8 @@ export {
   createConfig,
   findConfig,
   resolvePlugins,
+  clearPluginsCache,
+  getPluginCacheVersion,
   ConfigValidationError,
   Config, // FIXME: export it as a type
   type RawUniversalConfig,